Moves constants to template constants
