--- app/client/src/constants/TemplatesConstants.ts
@@ -1,7 +1,0 @@
-export const COMMUNITY_PORTAL = {
-  BASE_URL: "https://community.appsmith.com",
-};
-
-export const STARTER_BUILDING_BLOCKS = {
-  DATASOURCE_PROMPT_DELAY: 3000,
-};
